GHI-232 Enabling HTML print page for individual responses.
