+++ templates/resultspdf_rate.mustache
@@ -78,7 +78,7 @@
     <thead>
     <tr>
         {{#headers}}
-        <th class="header" style="text-align:{{align}}; width:{{width}};" scope="col">{{{text}}}</th>
+        <th class="header" style="text-align:{{align}};" width="{{width}}" scope="col">{{{text}}}</th>
         {{/headers}}
     </tr>
     </thead>
@@ -87,15 +87,15 @@
     {{#choicelabelrow}}
     <tr class="">
         {{#column1}}
-        <td class="cell" style="text-align:{{align}}; width:{{width}};" scope="col">{{{text}}}</td>
+        <td class="cell" style="text-align:{{align}};" width="{{width}}" scope="col">{{{text}}}</td>
         {{/column1}}
         {{#column2}}
-            <td class="cell" style="text-align:{{align}}; width:{{width}};" scope="col">
-            <table style="width:100%" cellpadding="2" cellspacing="0" border="1">
+            <td class="cell" style="text-align:{{align}};" width="{{width}}" scope="col">
+            <table width="100%" cellpadding="2" cellspacing="0" border="1">
                 <tbody>
                 <tr>
                     {{#ranks}}
-                        <td class="smalltext" style="text-align:center; width:{{width}};">{{{text}}}</td>
+                        <td class="smalltext" style="text-align:center;" width="{{width}}">{{{text}}}</td>
                     {{/ranks}}
                 </tr>
                 </tbody>
@@ -103,10 +103,10 @@
             </td>
         {{/column2}}
         {{#column3}}
-            <td class="cell" style="text-align:{{align}}; width:{{width}};" scope="col">{{{text}}}</td>
+            <td class="cell" style="text-align:{{align}};" width="{{width}}" scope="col">{{{text}}}</td>
         {{/column3}}
         {{#column4}}
-            <td class="cell" style="text-align:{{align}}; width:{{width}};" scope="col">{{{text}}}</td>
+            <td class="cell" style="text-align:{{align}};" width="{{width}}" scope="col">{{{text}}}</td>
         {{/column4}}
     </tr>
     {{/choicelabelrow}}
@@ -114,25 +114,25 @@
     {{#choiceaverages}}
     <tr class="">
         {{#column1}}
-            <td class="cell" style="text-align:{{align}}; width:{{width}};" scope="col">{{{text}}}</td>
+            <td class="cell" style="text-align:{{align}};" width="{{width}}" scope="col">{{{text}}}</td>
         {{/column1}}
         {{#column2}}
-            <td class="cell" style="text-align:{{align}}; width:{{width}};" scope="col">
-                {{#imageurl}}<img style="height:12px; width: 6px;{{#margin}} {{margin}};{{/margin}}" alt="" src="{{imageurl}}" />{{/imageurl}}
+            <td class="cell" style="text-align:{{align}};" width="{{width}}" scope="col">
+                {{#spacerimage}}<img style="height:12px; width: {{marginpdf}}in;" alt="" src="{{spacerimage}}" />{{/spacerimage}}{{#imageurl}}<img style="height:12px; width: 6px;" alt="" src="{{imageurl}}" />{{/imageurl}}
             </td>
         {{/column2}}
         {{#column3}}
-            <td class="cell" style="text-align:{{align}}; width:{{width}};" scope="col">{{{text}}}</td>
+            <td class="cell" style="text-align:{{align}};" width="{{width}}" scope="col">{{{text}}}</td>
         {{/column3}}
         {{#column4}}
-            <td class="cell" style="text-align:{{align}}; width:{{width}};" scope="col">{{{text}}}</td>
+            <td class="cell" style="text-align:{{align}};" width="{{width}}" scope="col">{{{text}}}</td>
         {{/column4}}
     </tr>
     {{/choiceaverages}}
     {{#nodata.0}}
         <tr class="">
             {{#nodata}}
-                <td class="cell" style="text-align:{{align}}; width:{{width}};" scope="col">{{text}}</td>
+                <td class="cell" style="text-align:{{align}};" width="{{width}}" scope="col">{{text}}</td>
             {{/nodata}}
         </tr>
     {{/nodata.0}}
